LayoutTests/imported/w3c/web-platform-tests/content-security-policy/wasm-unsafe-eval/default-src-blocks-wasm-streaming.any.worker-expected.txt


PASS WebAssembly.compileStreaming() is blocked
PASS WebAssembly.instantiateStreaming() is blocked
